core/src/test/java/com/avolution/actor/RemoteActorTest.java
package com.avolution.actor;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class RemoteActorTest {

    private RemoteActor remoteActor;

    @BeforeEach
    void setUp() {
        remoteActor = new RemoteActor("localhost", 8080);
    }

    @Test
    void testReceiveMessage() {
        Message message = new Message("Hello");
        remoteActor.receiveMessage(message);
        // Add assertions to verify message handling
    }

    @Test
    void testSendMessage() {
        RemoteActor recipient = new RemoteActor("localhost", 8081);
        Message message = new Message("Hello");
        remoteActor.sendMessage(recipient, message);
        // Add assertions to verify message sending
    }

    @Test
    void testRestart() {
        remoteActor.restart();
        // Add assertions to verify restart functionality
    }

    @Test
    void testStop() {
        remoteActor.stop();
        // Add assertions to verify stop functionality
    }

    @Test
    void testResume() {
        remoteActor.resume();
        // Add assertions to verify resume functionality
    }

    @Test
    void testIntegrationWithAbstractActor() {
        assertTrue(remoteActor instanceof AbstractActor);
    }
}
